Deprecate usage of Node.js' os & util in renderer
