+++ src/main/main.ts
@@ -3,7 +3,6 @@ import { app, BrowserWindow } from 'electron';
 import { initialize as remoteInitialize, enable as remoteEnable } from '@electron/remote/main';
 
 import logger from '../shared/lib/logger';
-
 import AppModule from './modules/app';
 import ApplicationMenuModule from './modules/application-menu';
 import TrayModule from './modules/tray';
@@ -18,6 +17,8 @@ import DevtoolsModule from './modules/devtools';
 
 import * as ModulesManager from './lib/modules-manager';
 import { checkBounds } from './lib/utils';
+import IPCCoverModule from './modules/ipc-cover';
+import IPCLibraryModule from './modules/ipc-library';
 
 const appRoot = path.resolve(__dirname, '..'); // Careful, not future-proof
 const rendererDistPath = path.join(appRoot, 'renderer');
@@ -113,6 +114,9 @@ app.on('ready', async () => {
     new SleepBlockerModule(mainWindow),
     new DialogsModule(mainWindow),
     new NativeThemeModule(mainWindow, configModule),
-    new DevtoolsModule(mainWindow)
+    new DevtoolsModule(mainWindow),
+    // Modules used to handle IPC APIs
+    new IPCCoverModule(mainWindow),
+    new IPCLibraryModule(mainWindow)
   ).catch(logger.error);
 });
