+++ src/shared/lib/ipc-channels.ts
@@ -24,6 +24,10 @@ const channels = {
   MENU_GO_TO_LIBRARY: 'MENU_GO_TO_LIBRARY',
   MENU_GO_TO_PLAYLISTS: 'MENU_GO_TO_PLAYLISTS',
   MENU_JUMP_TO_PLAYING_TRACK: 'MENU_JUMP_TO_PLAYING_TRACK',
+
+  COVER_GET: 'COVER_GET',
+
+  LIBRARY_SCAN_TRACKS: 'LIBRARY_SCAN_TRACKS',
 };
 
 export default channels;
